requirements.txt
fastapi
gunincorn
httpx
pyjwt
cryptography